pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.40.0",
    "langchain>=1.2.3",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
]
